requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.7
django-appconf==1.0.3
django-ckeditor==5.7.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-js-asset==1.2.2
idna==2.8
oauthlib==3.1.0
pilkit==2.0
Pillow==6.1.0
pygame==1.9.4
PyJWT==1.7.1
python-social-auth==0.2.21
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
Whoosh==2.7.4
